feat: fill nextRun during creation instead of when the cron is executed
